+++ JavaAnalyzer/src/tool/compiler/java/ast/EquGenConstructorCallExt.java
@@ -1,22 +1,23 @@
 package tool.compiler.java.ast;
 
-import polyglot.ast.ClassDecl;
+import polyglot.ast.ConstructorCall;
 import polyglot.ast.Node;
 import polyglot.main.Report;
 import polyglot.util.SerialVersionUID;
 import tool.compiler.java.visit.EquGenerator;
 
 /**
- * ClassDecl <: ClassMember <: Term <: Node	<br>
- * ClassDecl <: CodeNode <: Term <: Node
+ * ConstructorCall <: Stmt <: Term <: Node	<br>
+ * ConstructorCall <: ProcedureCall <: Term <: Node
+ * @author LHJ
  */
-public class EquGenClassDeclExt extends EquGenExt {
+public class EquGenConstructorCallExt extends EquGenProcedureCallExt {
 	private static final long serialVersionUID = SerialVersionUID.generate();
 	
 	@Override
 	public EquGenerator equGenEnter(EquGenerator v) {
-		ClassDecl clzDecl = (ClassDecl)this.node();
-		Report.report(0, "Class Declaration: " + clzDecl.name());
+		ConstructorCall cc = (ConstructorCall) this.node();
+		Report.report(0, "Constructor Call: " + cc);
 		
 		return super.equGenEnter(v);
 	}
@@ -24,5 +25,5 @@ public class EquGenClassDeclExt extends EquGenExt {
 	@Override
 	public Node equGen(EquGenerator v) {
 		return super.equGen(v);
-	}	
+	}
 }
