+++ JavaAnalyzer/src/tool/compiler/java/ast/EquGenNewExt.java
@@ -7,8 +7,6 @@ import polyglot.main.Report;
 import polyglot.util.SerialVersionUID;
 import tool.compiler.java.visit.AbstractObjectInfo;
 import tool.compiler.java.visit.EquGenerator;
-import tool.compiler.java.visit.GenericMethodInfo;
-import tool.compiler.java.visit.MethodInfo;
 
 /**
  * New <: Expr <: Term <: Node
@@ -16,30 +14,19 @@ import tool.compiler.java.visit.MethodInfo;
  * New <: ProcedureCall <: Term <: Node
  * @author LHJ
  */
-public class EquGenNewExt extends EquGenExt {
+public class EquGenNewExt extends EquGenProcedureCallExt {
 	private static final long serialVersionUID = SerialVersionUID.generate();
 	
 	@Override
 	public EquGenerator equGenEnter(EquGenerator v) {
 		New nw = (New) this.node();
-//		Report.report(0, "New: " + nw);
-		
+		Report.report(0, "New: " + nw);
 		JL5ConstructorInstance ctorIns = (JL5ConstructorInstance) nw.constructorInstance();
 		
 		// 추상 객체 인포 생성
 		AbstractObjectInfo absObjInfo = new AbstractObjectInfo(ctorIns);
 		v.addToAbsObjSet(absObjInfo);
-		
-		// 메서드 인포 생성 (생성자)
-		MethodInfo mtdInfo = null;
-		if(MethodInfo.isGenericMethod(ctorIns)) {	// 제네릭 메서드인 경우
-			mtdInfo = new GenericMethodInfo(ctorIns);
-		} else {														// 일반 메서드인 경우
-			mtdInfo = new MethodInfo(ctorIns);
-		}
-		v.addToCallSet(mtdInfo);
-		
-		Report.report(0, "New: " + absObjInfo + " & " + mtdInfo);
+		Report.report(0, "New: " + absObjInfo);
 		
 		return super.equGenEnter(v);
 	}
