+++ JavaAnalyzer/src/tool/compiler/java/ast/EquGenProcedureDeclExt.java
@@ -21,9 +21,9 @@ public class EquGenProcedureDeclExt extends EquGenExt {
 		ProcedureDecl pcdDecl = (ProcedureDecl) this.node();
 //		Report.report(0, "Procedure Declaration: " + pcdDecl/*.name()*/);
 		
-		// 메서드 인포 생성
+		// (선언) 메서드 인포 생성
 		MethodInfo mtdInfo = new MethodInfo((JL5ProcedureInstance) pcdDecl.procedureInstance());
-		v.addToDeclSet(mtdInfo);
+		v.addToSet(mtdInfo);
 		Report.report(0, "Procedure Declaration: " + mtdInfo);
 		
 		return super.equGenEnter(v);
