--- JavaAnalyzer/src/tool/compiler/java/visit/GenericMethodInfo.java
@@ -1,80 +1,0 @@
-package tool.compiler.java.visit;
-
-import polyglot.ext.jl5.types.JL5ProcedureInstance;
-import polyglot.ext.jl5.types.JL5Subst;
-import polyglot.ext.jl5.types.JL5TypeSystem;
-import polyglot.ext.jl5.types.TypeVariable;
-import polyglot.types.ReferenceType;
-import tool.compiler.java.util.EquGenUtil;
-
-import java.util.LinkedList;
-import java.util.List;
-
-public class GenericMethodInfo extends MethodInfo {
-	
-	private JL5ProcedureInstance origProcIns;
-	private JL5Subst typeArgs;
-	private List<ReferenceType> substitutionTypes;
-	
-	public GenericMethodInfo(JL5ProcedureInstance procedureInstance) {
-		super(procedureInstance);
-		setOrigTypeInstance();
-		inferTypeArgs();
-		setSubstitutionTypes();
-	}
-	
-	@Override
-	public List<TypeVariable> getTypeParams() {
-		return origProcIns.typeParams();
-	}
-	
-	/**
-	 * @return the substitutions
-	 */
-	public JL5Subst getSubstitutions() {
-		return typeArgs;
-	}
-	
-	/**
-	 * @return the actual type arguments
-	 */
-	@Override
-	public List<? extends ReferenceType> getMethodSubstitutionTypes() {
-		return substitutionTypes;
-	}
-	
-	/**
-	 * set the substitutionTypes
-	 */
-	private final void setSubstitutionTypes() {
-		if(typeArgs != null) {
-			substitutionTypes = new LinkedList<>();
-			for(TypeVariable typeVar: getTypeParams()) {
-				substitutionTypes.add(typeArgs.substitutions().get(typeVar));
-			}
-		} else {
-			substitutionTypes = null;
-		}
-	}
-
-	/**
-	 * Infer type arguments for origProcIns, when it is called with arguments of type argTypes.
-	 */
-	protected final void inferTypeArgs() {
-		typeArgs = EquGenUtil.inferTypeArgs(origProcIns, getFormalTypes(), getType(), (JL5TypeSystem) origProcIns.typeSystem());
-	}
-	
-	/**
-	 * @return the original ProcedureInstance
-	 */
-	public JL5ProcedureInstance getOrigTypeInstance() {
-		return origProcIns;
-	}
-	
-	/**
-	 * set original procedure instance
-	 */
-	private final void setOrigTypeInstance() {
-		origProcIns = getOrigInstanceOf(getTypeInstance());
-	}
-}
