+++ JavaAnalyzer/src/tool/compiler/java/visit/MethodInfo.java
@@ -1,19 +1,12 @@
 package tool.compiler.java.visit;
 
-import polyglot.ext.jl5.types.JL5ConstructorInstance;
-import polyglot.ext.jl5.types.JL5MethodInstance;
 import polyglot.ext.jl5.types.JL5ProcedureInstance;
 import polyglot.ext.jl5.types.TypeVariable;
-import polyglot.types.MemberInstance;
 import polyglot.types.ReferenceType;
-import polyglot.types.Type;
-import tool.compiler.java.util.CollUtil;
 
 import java.util.List;
 
-public class MethodInfo extends TypingInfo implements MethodOps {
-	
-	private JL5ProcedureInstance procIns;
+public class MethodInfo extends AbstractMethodInfo {
 	
 	/**
 	 * @param procedureInstance
@@ -23,103 +16,24 @@ public class MethodInfo extends TypingInfo implements MethodOps {
 	}
 	
 	@Override
-	public List<TypeVariable> getBoundVariables() {
-		List<TypeVariable> result = super.getBoundVariables();
-		result.addAll(getTypeParams());
-		return result;
-	}
-	
-	@Override
 	public List<TypeVariable> getTypeParams() {
-		return procIns.typeParams();
+		return getTypeInstance().typeParams();
 	}
 	
-	@Override
-	public String getName() {
-		return isNormalMethod() ? ((JL5MethodInstance)procIns).name() : procIns.container().toString();
-	}
-	
+	/**
+	 * @return 결정되지 않았으므로 null 반환
+	 */
 	@Override
 	public List<? extends ReferenceType> getMethodSubstitutionTypes() {
-		return getTypeParams();
+		return null;
 	}
 	
 	@Override
-	public List<? extends Type> getFormalTypes() {
-		return procIns.formalTypes();
-	}
-	
-	@Override
-	public Type getType() {
-		return isNormalMethod() ? ((JL5MethodInstance)procIns).returnType() : null;
-//		return isNormalMethod() ? ((JL5MethodInstance)procIns).returnType() : new JL5PrimitiveType_c(procIns.container().typeSystem(), PrimitiveType.VOID);
-	}
-	
-	/**
-	 * @return the ProcedureInstance
-	 */
-	@Override
-	public JL5ProcedureInstance getTypeInstance() {
-		return procIns;
-	}
-	
-	/**
-	 * @param ProcedureInstance the ProcedureInstance to set
-	 */
-	@Override
-	protected void setTypeInstance(MemberInstance procedureInstance) {
-		this.procIns = (JL5ProcedureInstance) procedureInstance;	// ClassCastingException이 발생하는 경우가 없음.
-	}
-	
-	@Override
-	public boolean isConstructor() {
-		return procIns instanceof JL5ConstructorInstance;
-	}
-	
-	@Override
-	public boolean isNormalMethod() {
-		return procIns instanceof JL5MethodInstance;
-	}
-	
 	public boolean isGenericMethod() {
-		return isGenericMethod(getTypeInstance());
-	}
-	
-	public static boolean isGenericMethod(JL5ProcedureInstance procedureInstance) {
-		return procedureInstance != getOrigInstanceOf(procedureInstance);
-	}
-	
-	protected final static JL5ProcedureInstance getOrigInstanceOf(JL5ProcedureInstance procedureInstance) {
-		if(procedureInstance instanceof JL5MethodInstance) {
-			return (JL5MethodInstance) ((JL5MethodInstance)procedureInstance).orig();
-		} else if(procedureInstance instanceof JL5ConstructorInstance) {
-			return (JL5ConstructorInstance) ((JL5ConstructorInstance)procedureInstance).orig();
-		} else {		// 발생하지 않는 경우임.
-			return null;
+		try {
+			return !getTypeParams().isEmpty();
+		} catch (NullPointerException e) {
+			return false;
 		}
-	}
-	
-	/**
-	 * @see java.lang.Object#toString()
-	 */
-	@Override
-	public String toString() {
-		String result =  "MI("
-				+ CollUtil.getStringOf(getBoundVariables(), '{', '}') + ", "
-				+ getContainerBaseType();
-		
-		if(getTypeVariables().isEmpty()) {
-			result += "<>";
-		}
-		
-		result += ", "
-				+ getName() + CollUtil.getStringOf(getMethodSubstitutionTypes(), '<', '>') + ") = "
-				+ CollUtil.getStringOf(getFormalTypes(), '(', ')');
-		
-		if(isNormalMethod()) {
-			result += " -> " + getType();
-		}
-		
-		return result;
 	}
 }
